FASE 2/Evidencias de proyecto/Evidencias de sistema/Aplicación/middleware_capstone/requirements.txt
apache-airflow==2.10.0
airflow-provider-great-expectations==0.2.8
great-expectations==0.18.19
apache-airflow-providers-common-sql
apache-airflow-providers-postgres==5.11.3
psycopg2-binary==2.9.9
pandas==2.1.4
pyarrow==17.0.0
python-decouple==3.8
python-dotenv==1.0.1